requirements.txt
openai-whisper>=20231117
anthropic>=0.18.0
numpy>=1.24.0
